+++ imap_tools/mailbox.py
@@ -53,7 +53,7 @@ class BaseMailBox:
 
     def search(self, criteria: str or bytes = 'ALL', charset: str = 'US-ASCII') -> [str]:
         """
-        Search mailbox for matching message numbers (this is not uids)
+        Search mailbox for matching message numbers in current folder (this is not uids)
         :param criteria: message search criteria (see examples at ./doc/imap_search_criteria.txt)
         :param charset: IANA charset, indicates charset of the strings that appear in the search criteria. See rfc2978
         :return list of str
@@ -201,7 +201,7 @@ class BaseMailBox:
             timezone = datetime.timezone(datetime.timedelta(hours=0))
         else:
             timezone = datetime.datetime.now().astimezone().tzinfo  # system timezone
-        cleaned_flags = clean_flags(flag_set)
+        cleaned_flags = clean_flags(flag_set or [])
         typ, dat = self.box.append(
             encode_folder(folder),  # noqa
             '({})'.format(' '.join(cleaned_flags)) if cleaned_flags else None,