+++ tests/test_folders.py
@@ -29,27 +29,43 @@ class FoldersTest(MailboxTestCase):
             # SET, GET
             mailbox.folder.set(mailbox.folder_test_base)
             self.assertEqual(mailbox.folder.get(), mailbox.folder_test_base)
+
             # CREATE
             mailbox.folder.create(mailbox.folder_test_new)
             folder_list_names = [i['name'] for i in mailbox.folder.list(mailbox.folder_test)]
             self.assertTrue(mailbox.folder_test_new in folder_list_names)
+
             # EXISTS
             self.assertTrue(mailbox.folder.exists(mailbox.folder_test_new))
+
             # RENAME
             mailbox.folder.rename(mailbox.folder_test_new, mailbox.folder_test_new1)
             folder_list_names = [i['name'] for i in mailbox.folder.list(mailbox.folder_test)]
             self.assertTrue(mailbox.folder_test_new1 in folder_list_names)
             self.assertFalse(mailbox.folder_test_new in folder_list_names)
+
             # DELETE
             mailbox.folder.delete(mailbox.folder_test_new1)
             folder_list_names = [i['name'] for i in mailbox.folder.list(mailbox.folder_test)]
             self.assertFalse(mailbox.folder_test_new1 in folder_list_names)
+
             # STATUS
             for status_key, status_val in mailbox.folder.status(mailbox.folder_test_base).items():
                 self.assertIs(type(status_key), str)
                 self.assertIs(type(status_val), int)
+
+            # SUBSCRIBE
+            if mailbox.mailbox_name not in ('MAIL_RU',):
+                mailbox.folder.subscribe(mailbox.folder_test_temp2, False)
+                self.assertIn(mailbox.folder_test_temp2, [i['name'] for i in mailbox.folder.list()])
+                self.assertNotIn(mailbox.folder_test_temp2,
+                                 [i['name'] for i in mailbox.folder.list(subscribed_only=1)])
+                mailbox.folder.subscribe(mailbox.folder_test_temp2, True)
+                self.assertIn(mailbox.folder_test_temp2,
+                              [i['name'] for i in mailbox.folder.list(subscribed_only=1)])
+
             # error
-            if mailbox.mailbox_name != 'YAHOO':
+            if mailbox.mailbox_name not in ('YAHOO',):
                 # YAHOO:
                 #   imaplib.IMAP4.error: SELECT command error: BAD [b'[TRYCREATE] SELECT error -
                 #   Folder does not exist or server encountered an error']